--- addons/rebalance/CfgVehicles.hpp
@@ -1,9 +1,0 @@
-class CfgVehicles {
-    class Bag_Base;
-    class fatpack_od: Bag_Base {
-        maximumLoad = 180;
-    };
-    class fatpack_coy: fatpack_od {
-        maximumLoad = 180;
-    };
-};
