--- addons/rebalance/CfgWeapons.hpp
@@ -1,11 +1,0 @@
-class CfgWeapons {
-    class ItemCore;
-    class VestItem;
-    class Vest_Camo_Base;
-
-    #include "milgp_jpc.hpp"
-    #include "milgp_mmac.hpp"
-    #include "milgp_marciras.hpp"
-    #include "s4_cpc.hpp"
-    #include "s4_lbt.hpp"
-};
