+++ addons/rebalance/milgp/CfgWeapons.hpp
@@ -1,0 +1,477 @@
+class CfgWeapons {
+    // JPC
+    class VestItem;
+    class ItemCore;
+    class milgp_v_jpc_assaulter_rgr: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_assaulter_mc: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_assaulter_cb: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_assaulter_khk: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_assaulter_belt_rgr: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_assaulter_belt_mc: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_assaulter_belt_cb: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_assaulter_belt_khk: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_Grenadier_rgr: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_Grenadier_mc: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_Grenadier_cb: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_Grenadier_khk: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_Grenadier_belt_rgr: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_Grenadier_belt_mc: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_Grenadier_belt_cb: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_Grenadier_belt_khk: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_hgunner_rgr: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_hgunner_mc: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_hgunner_cb: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_hgunner_khk: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_hgunner_belt_rgr: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_hgunner_belt_mc: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_hgunner_belt_cb: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_hgunner_belt_khk: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_Medic_rgr: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_Medic_mc: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_Medic_cb: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_Medic_khk: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_medic_belt_rgr: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_medic_belt_mc: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_medic_belt_cb: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_medic_belt_khk: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_Marksman_rgr: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_Marksman_mc: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_Marksman_cb: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_Marksman_khk: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_marksman_belt_rgr: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_marksman_belt_mc: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_marksman_belt_cb: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_marksman_belt_khk: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_TeamLeader_rgr: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_TeamLeader_mc: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_TeamLeader_cb: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_TeamLeader_khk: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_teamleader_belt_rgr: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_teamleader_belt_mc: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_teamleader_belt_cb: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_teamleader_belt_khk: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_Light_rgr: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_Light_mc: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_Light_cb: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+    class milgp_v_jpc_Light_khk: ItemCore {
+        MACRO_HITPROTECTION_JPC;
+    };
+
+    // Marciras
+    class milgp_v_marciras_teamleader_mc: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_teamleader_belt_mc: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_assaulter_mc: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_assaulter_belt_mc: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_grenadier_mc: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_grenadier_belt_mc: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_marksman_mc: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_marksman_belt_mc: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_hgunner_mc: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_hgunner_belt_mc: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_medic_mc: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_medic_belt_mc: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_light_mc: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_teamleader_RGR: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_teamleader_belt_RGR: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_assaulter_rgr: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_assaulter_belt_rgr: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_grenadier_rgr: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_grenadier_belt_rgr: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_marksman_rgr: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_marksman_belt_rgr: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_hgunner_rgr: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_hgunner_belt_rgr: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_medic_rgr: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_medic_belt_rgr: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_light_rgr: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_teamleader_CB: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_teamleader_belt_CB: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_assaulter_cb: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_assaulter_belt_cb: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_grenadier_cb: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_grenadier_belt_cb: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_marksman_cb: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_marksman_belt_cb: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_hgunner_cb: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_hgunner_belt_cb: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_medic_cb: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_medic_belt_cb: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_light_cb: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_teamleader_KHK: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_teamleader_belt_KHK: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_assaulter_khk: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_assaulter_belt_khk: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_grenadier_khk: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_grenadier_belt_khk: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_marksman_khk: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_marksman_belt_khk: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_hgunner_khk: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_hgunner_belt_khk: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_medic_khk: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_medic_belt_khk: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+    class milgp_v_marciras_light_khk: ItemCore {
+        MACRO_HITPROTECTION_MARCIRAS;
+    };
+
+    // MMAC
+    class milgp_v_mmac_light_rgr: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_assaulter_rgr: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_assaulter_belt_rgr: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_medic_rgr: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_medic_belt_rgr: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_teamleader_rgr: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_teamleader_belt_rgr: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_marksman_rgr: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_marksman_belt_rgr: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_hgunner_rgr: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_hgunner_belt_rgr: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_grenadier_rgr: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_grenadier_belt_rgr: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_light_KHK: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_assaulter_KHK: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_assaulter_belt_KHK: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_medic_KHK: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_medic_belt_KHK: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_teamleader_KHK: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_teamleader_belt_KHK: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_marksman_KHK: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_marksman_belt_KHK: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_hgunner_KHK: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_hgunner_belt_KHK: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_grenadier_KHK: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_grenadier_belt_KHK: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_light_MC: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_assaulter_MC: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_assaulter_belt_MC: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_medic_MC: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_medic_belt_MC: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_teamleader_MC: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_teamleader_belt_MC: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_marksman_MC: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_marksman_belt_MC: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_hgunner_MC: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_hgunner_belt_MC: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_grenadier_MC: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_grenadier_belt_MC: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_light_CB: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_assaulter_CB: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_assaulter_belt_CB: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_medic_CB: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_medic_belt_CB: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_teamleader_CB: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_teamleader_belt_CB: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_marksman_CB: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_marksman_belt_CB: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_hgunner_CB: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_hgunner_belt_CB: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_grenadier_CB: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+    class milgp_v_mmac_grenadier_belt_CB: ItemCore {
+        MACRO_HITPROTECTION_MMAC;
+    };
+};
