--- addons/rebalance/milgp_jpc.hpp
@@ -1,182 +1,0 @@
-    #define MACRO_HITPROTECTION_JPC \
-        class ItemInfo: VestItem { \
-            class HitpointsProtectionInfo { \
-                class Chest { \
-                    HitpointName = "HitChest"; \
-                    armor = 15; \
-                    passThrough = 0.45; \
-                }; \
-                class Diaphragm { \
-                    HitpointName = "HitDiaphragm"; \
-                    armor = 15; \
-                    passThrough = 0.45; \
-                }; \
-                class Abdomen { \
-                    hitpointName = "HitAbdomen"; \
-                    armor = 15; \
-                    passThrough = 0.45; \
-                }; \
-                class Body { \
-                    hitpointName = "HitBody"; \
-                    armor = 15; \
-                    passThrough = 0.45; \
-                }; \
-            }; \
-        };
-
-    class milgp_v_jpc_assaulter_rgr: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_assaulter_mc: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_assaulter_cb: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_assaulter_khk: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_assaulter_belt_rgr: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_assaulter_belt_mc: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_assaulter_belt_cb: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_assaulter_belt_khk: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_Grenadier_rgr: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_Grenadier_mc: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_Grenadier_cb: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_Grenadier_khk: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_Grenadier_belt_rgr: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_Grenadier_belt_mc: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_Grenadier_belt_cb: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_Grenadier_belt_khk: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_hgunner_rgr: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_hgunner_mc: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_hgunner_cb: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_hgunner_khk: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_hgunner_belt_rgr: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_hgunner_belt_mc: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_hgunner_belt_cb: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_hgunner_belt_khk: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_Medic_rgr: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_Medic_mc: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_Medic_cb: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_Medic_khk: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_medic_belt_rgr: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_medic_belt_mc: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_medic_belt_cb: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_medic_belt_khk: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_Marksman_rgr: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_Marksman_mc: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_Marksman_cb: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_Marksman_khk: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_marksman_belt_rgr: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_marksman_belt_mc: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_marksman_belt_cb: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_marksman_belt_khk: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_TeamLeader_rgr: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_TeamLeader_mc: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_TeamLeader_cb: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_TeamLeader_khk: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_teamleader_belt_rgr: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_teamleader_belt_mc: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_teamleader_belt_cb: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_teamleader_belt_khk: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_Light_rgr: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_Light_mc: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_Light_cb: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
-    class milgp_v_jpc_Light_khk: ItemCore {
-        MACRO_HITPROTECTION_JPC
-    };
