--- addons/rebalance/milgp_marciras.hpp
@@ -1,182 +1,0 @@
-#define MACRO_HITPROTECTION_MARCIRAS \
-        class ItemInfo: VestItem { \
-            class HitpointsProtectionInfo { \
-                class Chest { \
-                    HitpointName = "HitChest"; \
-                    armor = 18; \
-                    passThrough = 0.3; \
-                }; \
-                class Diaphragm { \
-                    HitpointName = "HitDiaphragm"; \
-                    armor = 18; \
-                    passThrough = 0.3; \
-                }; \
-                class Abdomen { \
-                    hitpointName = "HitAbdomen"; \
-                    armor = 18; \
-                    passThrough = 0.3; \
-                }; \
-                class Body { \
-                    hitpointName = "HitBody"; \
-                    armor = 18; \
-                    passThrough = 0.3; \
-                }; \
-            }; \
-        };
-
-class milgp_v_marciras_teamleader_mc: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_teamleader_belt_mc: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_assaulter_mc: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_assaulter_belt_mc: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_grenadier_mc: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_grenadier_belt_mc: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_marksman_mc: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_marksman_belt_mc: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_hgunner_mc: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_hgunner_belt_mc: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_medic_mc: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_medic_belt_mc: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_light_mc: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_teamleader_RGR: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_teamleader_belt_RGR: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_assaulter_rgr: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_assaulter_belt_rgr: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_grenadier_rgr: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_grenadier_belt_rgr: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_marksman_rgr: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_marksman_belt_rgr: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_hgunner_rgr: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_hgunner_belt_rgr: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_medic_rgr: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_medic_belt_rgr: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_light_rgr: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_teamleader_CB: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_teamleader_belt_CB: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_assaulter_cb: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_assaulter_belt_cb: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_grenadier_cb: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_grenadier_belt_cb: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_marksman_cb: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_marksman_belt_cb: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_hgunner_cb: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_hgunner_belt_cb: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_medic_cb: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_medic_belt_cb: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_light_cb: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_teamleader_KHK: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_teamleader_belt_KHK: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_assaulter_khk: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_assaulter_belt_khk: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_grenadier_khk: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_grenadier_belt_khk: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_marksman_khk: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_marksman_belt_khk: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_hgunner_khk: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_hgunner_belt_khk: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_medic_khk: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_medic_belt_khk: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
-class milgp_v_marciras_light_khk: ItemCore {
-    MACRO_HITPROTECTION_MARCIRAS
-};
