--- addons/rebalance/milgp_mmac.hpp
@@ -1,182 +1,0 @@
-#define MACRO_HITPROTECTION_MMAC \
-        class ItemInfo: VestItem { \
-            class HitpointsProtectionInfo { \
-                class Chest { \
-                    HitpointName = "HitChest"; \
-                    armor = 16; \
-                    passThrough = 0.4; \
-                }; \
-                class Diaphragm { \
-                    HitpointName = "HitDiaphragm"; \
-                    armor = 16; \
-                    passThrough = 0.4; \
-                }; \
-                class Abdomen { \
-                    hitpointName = "HitAbdomen"; \
-                    armor = 16; \
-                    passThrough = 0.4; \
-                }; \
-                class Body { \
-                    hitpointName = "HitBody"; \
-                    armor = 16; \
-                    passThrough = 0.4; \
-                }; \
-            }; \
-        };
-
-class milgp_v_mmac_light_rgr: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_assaulter_rgr: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_assaulter_belt_rgr: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_medic_rgr: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_medic_belt_rgr: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_teamleader_rgr: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_teamleader_belt_rgr: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_marksman_rgr: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_marksman_belt_rgr: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_hgunner_rgr: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_hgunner_belt_rgr: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_grenadier_rgr: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_grenadier_belt_rgr: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_light_KHK: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_assaulter_KHK: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_assaulter_belt_KHK: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_medic_KHK: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_medic_belt_KHK: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_teamleader_KHK: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_teamleader_belt_KHK: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_marksman_KHK: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_marksman_belt_KHK: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_hgunner_KHK: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_hgunner_belt_KHK: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_grenadier_KHK: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_grenadier_belt_KHK: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_light_MC: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_assaulter_MC: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_assaulter_belt_MC: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_medic_MC: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_medic_belt_MC: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_teamleader_MC: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_teamleader_belt_MC: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_marksman_MC: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_marksman_belt_MC: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_hgunner_MC: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_hgunner_belt_MC: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_grenadier_MC: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_grenadier_belt_MC: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_light_CB: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_assaulter_CB: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_assaulter_belt_CB: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_medic_CB: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_medic_belt_CB: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_teamleader_CB: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_teamleader_belt_CB: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_marksman_CB: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_marksman_belt_CB: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_hgunner_CB: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_hgunner_belt_CB: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_grenadier_CB: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
-class milgp_v_mmac_grenadier_belt_CB: ItemCore {
-    MACRO_HITPROTECTION_MMAC
-};
