+++ addons/rebalance/s4g/CfgWeapons.hpp
@@ -1,0 +1,119 @@
+class CfgWeapons {
+    // CPC
+    class VestItem;
+    class Vest_Camo_Base;
+    class cpc_Fast_coy: Vest_Camo_Base {
+        MACRO_HITPROTECTION_CPC;
+    };
+    class cpc_Fast_rngr: cpc_Fast_coy {};
+    class cpc_Fast_mc: cpc_Fast_coy {};
+
+    class cpc_Fastbelt_coy: Vest_Camo_Base {
+        MACRO_HITPROTECTION_CPC;
+    };
+    class cpc_Fastbelt_rngr: cpc_Fastbelt_coy {};
+    class cpc_Fastbelt_mc: cpc_Fastbelt_coy {};
+
+    class cpc_weaponsbelt_coy: Vest_Camo_Base {
+        MACRO_HITPROTECTION_CPC;
+    };
+    class cpc_weaponsbelt_mc: cpc_weaponsbelt_coy {};
+    class cpc_weaponsbelt_rngr: cpc_weaponsbelt_coy {};
+
+    class cpc_light_coy: Vest_Camo_Base {
+        MACRO_HITPROTECTION_CPC;
+    };
+    class cpc_light_rngr: cpc_light_coy {};
+    class cpc_light_mc: cpc_light_coy {};
+
+    class cpc_lightbelt_coy: Vest_Camo_Base {
+        MACRO_HITPROTECTION_CPC;
+    };
+    class cpc_lightbelt_rngr: cpc_lightbelt_coy {};
+    class cpc_lightbelt_mc: cpc_lightbelt_coy {};
+
+    class cpc_weapons_coy: Vest_Camo_Base {
+        MACRO_HITPROTECTION_CPC;
+    };
+    class cpc_weapons_rngr: cpc_weapons_coy {};
+    class cpc_weapons_mc: cpc_weapons_coy {};
+
+    class cpc_communications_coy: Vest_Camo_Base {
+        MACRO_HITPROTECTION_CPC;
+    };
+    class cpc_communications_mc: cpc_communications_coy {};
+    class cpc_communications_rngr: cpc_communications_coy {};
+
+    class cpc_communicationsbelt_coy: Vest_Camo_Base {
+        MACRO_HITPROTECTION_CPC;
+    };
+    class cpc_communicationsbelt_mc: cpc_communicationsbelt_coy {};
+    class cpc_communicationsbelt_rngr: cpc_communicationsbelt_coy {};
+
+    class cpc_tl_coy: Vest_Camo_Base {
+        MACRO_HITPROTECTION_CPC;
+    };
+    class cpc_tl_mc: cpc_tl_coy {};
+    class cpc_tl_rngr: cpc_tl_coy {};
+
+    class cpc_tlbelt_coy: Vest_Camo_Base {
+        MACRO_HITPROTECTION_CPC;
+    };
+    class cpc_tlbelt_rngr: cpc_tlbelt_coy {};
+    class cpc_tlbelt_mc: cpc_tlbelt_coy {};
+
+    class cpc_medical_coy: Vest_Camo_Base {
+        MACRO_HITPROTECTION_CPC;
+    };
+    class cpc_medical_mc: cpc_tl_coy {};
+    class cpc_medical_rngr: cpc_tl_coy {};
+
+    class cpc_medicalbelt_coy: Vest_Camo_Base {
+        MACRO_HITPROTECTION_CPC;
+    };
+    class cpc_medicalbelt_rngr: cpc_medicalbelt_coy {};
+    class cpc_medicalbelt_mc: cpc_medicalbelt_coy {};
+
+    // LBT
+    class lbt_pouchless_coy: Vest_Camo_Base {
+        MACRO_HITPROTECTION_LBT;
+    };
+    class lbt_pouchless_od: lbt_pouchless_coy {};
+    class lbt_pouchless_aor1: lbt_pouchless_coy {};
+    class lbt_pouchless_aor2: lbt_pouchless_coy {};
+    class lbt_pouchless_mc: lbt_pouchless_coy {};
+    class lbt_operator_coy: lbt_pouchless_coy {};
+    class lbt_operator_od: lbt_pouchless_coy {};
+    class lbt_operator_aor1: lbt_pouchless_coy {};
+    class lbt_operator_aor2: lbt_pouchless_coy {};
+    class lbt_operator_mc: lbt_pouchless_coy {};
+    class lbt_weapons_coy: lbt_pouchless_coy {};
+    class lbt_weapons_od: lbt_weapons_coy {};
+    class lbt_weapons_aor1: lbt_weapons_coy {};
+    class lbt_weapons_aor2: lbt_weapons_coy {};
+    class lbt_weapons_mc: lbt_pouchless_coy {};
+    class lbt_medical_coy: lbt_pouchless_coy {};
+    class lbt_medical_od: lbt_medical_coy {};
+    class lbt_medical_aor1: lbt_medical_coy {};
+    class lbt_medical_aor2: lbt_medical_coy {};
+    class lbt_medical_mc: lbt_pouchless_coy {};
+    class lbt_comms_coy: lbt_pouchless_coy {};
+    class lbt_comms_od: lbt_comms_coy {};
+    class lbt_comms_aor1: lbt_comms_coy {};
+    class lbt_comms_aor2: lbt_comms_coy {};
+    class lbt_comms_mc: lbt_pouchless_coy {};
+    class lbt_light_od: lbt_pouchless_coy {};
+    class lbt_light_mc: lbt_pouchless_coy {};
+    class lbt_light_aor1: lbt_pouchless_coy {};
+    class lbt_light_aor2: lbt_pouchless_coy {};
+    class lbt_light_coy: lbt_pouchless_coy {};
+    class lbt_fast_coy: lbt_pouchless_coy {};
+
+    class lbt_tl_mc: Vest_Camo_Base {
+        MACRO_HITPROTECTION_LBT;
+    };
+    class lbt_tl_coy: lbt_tl_mc {};
+    class lbt_tl_od: lbt_tl_mc {};
+    class lbt_tl_aor1: lbt_tl_mc {};
+    class lbt_tl_aor2: lbt_tl_mc {};
+};
